+++ loader/src/ui/mods/list/ModItem.cpp
@@ -140,7 +140,7 @@ bool ModItem::init(ModSource&& source) {
 
         auto geodeValid = Loader::get()->isModVersionSupported(version.getGeodeVersion());
         auto gameVersion = version.getGameVersion();
-        auto gdValid = gameVersion == "*" || gameVersion == GEODE_STR(GEODE_GD_VERSION);
+        auto gdValid = !gameVersion || gameVersion == "*" || gameVersion == GEODE_STR(GEODE_GD_VERSION);
 
         if (!geodeValid || !gdValid) {
             spr = createGeodeButton("N/A", 50, false, true, GeodeButtonSprite::Gray);
@@ -496,21 +496,28 @@ void ModItem::onView(CCObject*) {
     // Show popups for invalid mods
     if (m_source.asServer()) {
         auto version = m_source.asServer()->latestVersion();
-        if (!Loader::get()->isModVersionSupported(version.getGeodeVersion())) {
+        auto gameVersion = version.getGameVersion();
+        if (gameVersion == "0.000") {
             return FLAlertLayer::create(
                 nullptr,
-                "Outdated",
-                "This mod is targets an <cr>outdated version of Geode</c>. "
-                "<co>Please wait for its developer to update it.</c>",
+                "Invalid Platform",
+                "This mod is <cr>not available</c> for your current platform.",
                 "OK", nullptr, 360
             )->show();
         }
-        if (version.getGameVersion() != "*" && version.getGameVersion() != GEODE_STR(GEODE_GD_VERSION)) {
+        if (gameVersion && gameVersion != "*" && gameVersion != GEODE_STR(GEODE_GD_VERSION)) {
             return FLAlertLayer::create(
                 nullptr,
-                "Outdated",
-                "This mod is targets a <cr>different version of Geometry Dash</c>. "
-                "<co>Please wait for its developer to update it.</c>",
+                "Unavailable",
+                "This mod targets an <cr>unsupported version of Geometry Dash</c>.",
+                "OK", nullptr, 360
+            )->show();
+        }
+        if (!Loader::get()->isModVersionSupported(version.getGeodeVersion())) {
+            return FLAlertLayer::create(
+                nullptr,
+                "Unavailable",
+                "This mod targets an <cr>unsupported version of Geode</c>.",
                 "OK", nullptr, 360
             )->show();
         }